# main.cpp ---
* Added more cleanup code

# IOManager.cpp
---

* Added License notice to TOF

# IOManager.h
---

* Added License notice to TOF

# FileManager
---

* Added a dedicated shutdown function and replaced the code within the deconstructor to just use the new function.

## FileJanitor
---

* Added a dedicated shutdown function and replaced the code within the deconstructor to just use the new function.

* Added generateEvictionChecklist() to replace the scoped code within evictUnused() that creates a checklist.

* Added hasExpired() to further scope the logic that checks if the file has outlived it's TTL.

### evictUnused()
---

* Added std::string cRecordFilePath to replace the ``fileRecord->path.string()`` calls, improving performance.

* Made some variables that didnt change const and renamed them accordingly.

* Renamed `last` to `cLastUseTime` as it is a better name.

* Added a few comments.

* Renamed `sp` when checking if the weakDataPtr of a FileRecord object is still valid to `recordDataPtr` as its a better name.

* Renamed `rec` to `fileRecord` to clarify what it is

* Moved the `cNow` variable closer to the scope of where it is used.

* Added error logging to notify if a file failed to evict

* Otherwise cleaned up this function

# EFormat
---

* Renamed to EPixelFormat

# Swapchain
---

* Removed all old code

* Added <vulkan/vulkan_core.h> include

* Added "EFormat.h" include

# VkDevice.h
---

* Added DescriptorAllocatorGrowable

# Core.h
---

* Greatly improved the TOF file documentation.

* Added a license statement at the TOF.

* Added documentation to some of the various macros within the file

* Added a few comments around the file

* Changed the unknown platform error message

# VulkanInstanceWrapper
---

* Fixed a bug where an error would get thrown on shutdown. Apparently, the destroy function was called twice for some reason

+++ src/engine/graphics/vulkan/VulkanInstance.cpp
@@ -67,13 +67,30 @@ uint16_t Atlas::VulkanInstanceWrapper::init() {
 	mVulkanInstance = instanceReturn.value().instance;
 	mDebugMessenger = instanceReturn.value().debug_messenger;
 
+	mbIsInitialized = true;
 	return initResult;
 }
 
 void Atlas::VulkanInstanceWrapper::shutdown()
 {
-	vkb::destroy_debug_utils_messenger(mVulkanInstance, mDebugMessenger);
+	if (!mbIsInitialized) {
+		WarnLog("Attempted to shut down a Vulkan instance that was not initialized or already shut down.");
+		return;
+	}
+
+	if (mVulkanInstance == VK_NULL_HANDLE)
+	{
+		WarnLog("Attempted to shut down a Vulkan instance whose instance handle is null.");
+		return;
+	}
+
+	if (mDebugMessenger != VK_NULL_HANDLE) {
+		vkb::destroy_debug_utils_messenger(mVulkanInstance, mDebugMessenger);
+	}
+
 	vkDestroyInstance(mVulkanInstance, nullptr);
+
+	mbIsInitialized = false;
 }
 
 VkInstance& Atlas::VulkanInstanceWrapper::getInstance() {
